refactor(backend): reorganize into module
