+++ backend/requirements.in
@@ -3,5 +3,6 @@ uvicorn[standard]
 python-multipart
 psycopg2
 typing_extensions
+pydantic ~= 2.0
 
 argon2-cffi~=23.1
