+++ backend/rotini/auth/base.py
@@ -1,0 +1,8 @@
+import pydantic
+
+
+class LoginRequestData(pydantic.BaseModel):
+    """Payload for login requests"""
+
+    username: str
+    password: str
