+++ backend/rotini/main.py
@@ -4,8 +4,8 @@ Rotini: a self-hosted cloud storage & productivity app.
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-import api.users
-import api.files
+import auth.routes as auth_routes
+import api.files as files_routes
 
 app = FastAPI()
 
@@ -19,8 +19,10 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-app.include_router(api.files.router)
-app.include_router(api.users.router)
+routers = [files_routes.router, auth_routes.router]
+
+for router in routers:
+    app.include_router(router)
 
 
 @app.get("/", status_code=204)
